requirements.txt
transformers==4.44.2
torch==2.4.1
langchain-groq==0.2.0
python-dotenv==1.0.1
gtts==2.5.3
streamlit==1.38.0
langchain==0.3.1
langchain-community==0.3.1
google-search-results==2.4.2
googletrans==4.0.0-rc1